made some updates
