+++ TODO.md
@@ -2,7 +2,12 @@
 [x] - added a todo file to keep track of things
 [x] - added the container,screens and font family options to the tailwind config and index.html file
 [x] - deleted the app.css file
-[] - ask team members to use uppercase letter for the components and store them on one component folder instead of different folders
+[x] - ask team members to use uppercase letter for the components and store them on one component folder instead of different folders
 [x] - redid the hero section
 [x] - made some changes to the search bar
 [x] - added a nav bar component
+[] - restructure the files/folders in order for it to be easy to navigate
+[] - complete the feature calender section
+[]- create a constants file to store the static data
+[] - get the banner image from cto
+[]
