+++ src/components/dashboard/index.jsx
@@ -1,13 +1,12 @@
 import React from 'react';
-import Header from './header';
-import Events from './events/index'
+import EventHeader from './EventHeader';
+import Events from './events/index';
 
 export default function index() {
   return (
     <div>
-        <Header/>
-        <Events/>
-
+      <EventHeader />
+      <Events />
     </div>
-  )
+  );
 }
