+++ src/libs/constants.js
@@ -1,0 +1,58 @@
+import Num from '../assets/Future100.png';
+import Rave from '../assets/rave.png';
+import Even from '../assets/123.png';
+import Chilbilz from '../assets/873.png';
+import Techbro from '../assets/tech.png';
+import David from '../assets/david.png';
+
+// data for the featured calender
+export const cards = [
+  {
+    id: 1,
+    name: 'Future100',
+    description: 'Join the community that revolves around tech.',
+    imgUrl: Num,
+  },
+  {
+    id: 2,
+    name: 'Rave',
+    description: 'Join the community that revolves around tech.',
+    imgUrl: Rave,
+  },
+  {
+    id: 3,
+    name: 'ChilBiz',
+    description: 'Join the community that revolves around tech.',
+    imgUrl: Even,
+  },
+  {
+    id: 4,
+    name: 'TechBro',
+    description: 'Join the community that revolves around tech.',
+    imgUrl: Chilbilz,
+  },
+  {
+    id: 4,
+    name: 'TechBro',
+    description: 'Join the community that revolves around tech.',
+    imgUrl: Techbro,
+  },
+  {
+    id: 5,
+    name: 'TechBro',
+    description: 'Join the community that revolves around tech.',
+    imgUrl: David,
+  },
+  {
+    id: 6,
+    name: 'TechBro',
+    description: 'Join the community that revolves around tech.',
+    imgUrl: David, // Replace with actual image URL
+  },
+  {
+    id: 7,
+    name: 'TechBro',
+    description: 'Join the community that revolves around tech.',
+    imgUrl: David, // Replace with actual image URL
+  },
+];
